Re-did it with a single server-file
